requirements.txt
# requirements.txt (FINAL CLEANED VERSION)
pandas==2.2.2
prophet==1.1.5
numpy==1.26.4
cmdstanpy==1.2.2
fastapi[all]
python-dotenv
uvicorn[standard]
kafka-python
mysql-connector-python
passlib==1.7.4
bcrypt==3.2.0
python-jose[cryptography]
cryptography==42.0.5
mlxtend==0.23.0
scikit-learn==1.4.2